Move definition of LaunchConfig class to separate file
This is necessary to avoid circular dependency.

Cluster-related occupancy functions need LaunchConfig.
Occupancy functions are defined in _module.py, and _launcher.py
that used to house definition of LaunchConfig imports Kernel
from _module.py
